requirements.txt
numpy>=1.24
opencv-python>=4.8
PyYAML>=6.0
matplotlib>=3.7
torch>=2.0
torchvision>=0.15
tqdm>=4.66
Pillow>=9.5
scipy>=1.10
pycocotools